refactor(schemas): move RBAC tables to schemas/rbac/
Per-dialect Drizzle table definitions for roles, permissions,
rolePermissions, userRoles, roleInherits, and userPermissionCache now live
in schemas/rbac/{postgres,mysql,sqlite}.ts with a dialect-aware barrel
exposing rbacTables(dialect).

database/schema/<dialect>.ts re-exports from the new location so existing
imports and relations() blocks keep compiling. Relations themselves stay
in database/schema/<dialect>.ts during Plan A and consolidate in Task 17.

Note: distinct from schemas/_zod/rbac.ts which holds the Zod validators for
the RBAC domain.
